removed legacy projects updated project.json to match original nuspec
