--- Semver.Core/project.json
@@ -1,15 +1,0 @@
-﻿{
-  "version": "1.0.0-*",
-
-  "dependencies": {
-    "NETStandard.Library": "1.5.0-rc2-24027"
-  },
-
-  "frameworks": {
-    "netstandard1": {
-      "imports": "dnxcore50"
-    },
-    "net20" : {
-    }
-  }
-}
