--- Semver/Semver.nuspec
@@ -1,23 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>semver</id>
-    <version>1.1.1</version>
-    <title>Semantic versioning</title>
-    <authors>Max Hauser</authors>
-    <owners>Max Hauser</owners>
-    <projectUrl>https://github.com/maxhauser/semver/</projectUrl>
-    <licenseUrl>http://max.mit-license.org/</licenseUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>A semver implementation in .Net based on the v2.0.0 of the spec found at http://semver.org/.</description>
-    <copyright>Copyright 2013</copyright>
-    <tags>semver semantic version</tags>
-  </metadata>
-  <files>
-      <file src="$outputfolder$\v2.0\semver.dll" target="lib\net20" />
-      <file src="$outputfolder$\v3.5\semver.dll" target="lib\net35" />
-      <file src="$outputfolder$\v4.0\semver.dll" target="lib\net40" />
-      <file src="$outputfolder$\v4.5\semver.dll" target="lib\net45" />
-      <file src="$outputfolder$\v4.5.1\semver.dll" target="lib\net451" />
-  </files>
-</package>
